% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensic_integration.R
\name{calculate_data_completeness}
\alias{calculate_data_completeness}
\title{Calculate Data Completeness}
\usage{
calculate_data_completeness(forensic_result)
}
\arguments{
\item{forensic_result}{IPVForensicResult object}
}
\value{
Completeness score (0-1)
}
\description{
Calculates overall data completeness score for forensic analysis.
}
