% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Generate Confusion Matrix}
\usage{
confusion_matrix(predictions, actual)
}
\arguments{
\item{predictions}{Predictions vector}

\item{actual}{Actual values vector}
}
\value{
Confusion matrix
}
\description{
Generate Confusion Matrix
}
