% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_interface.R
\name{create_error_response}
\alias{create_error_response}
\title{Create Error Response (Helper)}
\usage{
create_error_response(error)
}
\arguments{
\item{error}{Error object}
}
\value{
Error response structure
}
\description{
Create Error Response (Helper)
}
