% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensic_data_structure.R
\name{create_forensic_collection}
\alias{create_forensic_collection}
\title{Create IPV Forensic Analysis Collection}
\usage{
create_forensic_collection(incident_ids, data_source = NULL)
}
\arguments{
\item{incident_ids}{Vector of incident IDs}

\item{data_source}{Data source (tibble or data.frame) with narratives}
}
\value{
IPVForensicCollection object
}
\description{
Factory function to create and manage multiple IPVForensicResult instances
for batch analysis operations.
}
