% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{handle_na}
\alias{handle_na}
\title{Handle NA values consistently}
\usage{
handle_na(value, default = NA)
}
\arguments{
\item{value}{Value to check}

\item{default}{Default value if NA}
}
\value{
Value or default
}
\description{
Handle NA values consistently
}
\keyword{internal}
