% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensic_integration.R
\name{populate_from_batch_results}
\alias{populate_from_batch_results}
\title{Populate from Batch Results}
\usage{
populate_from_batch_results(forensic_result, basic_row, data_row)
}
\arguments{
\item{forensic_result}{IPVForensicResult object}

\item{basic_row}{Row from basic batch results}

\item{data_row}{Row from original data}
}
\value{
Updated forensic result
}
\description{
Populates forensic result with data from batch processing results.
}
