% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensic_data_structure.R
\name{print.IPVForensicCollection}
\alias{print.IPVForensicCollection}
\title{Print Method for IPV Forensic Collection}
\usage{
\method{print}{IPVForensicCollection}(x, ...)
}
\description{
Print Method for IPV Forensic Collection
}
