% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensic_data_structure.R
\name{print.IPVForensicResult}
\alias{print.IPVForensicResult}
\title{Print Method for IPV Forensic Result}
\usage{
\method{print}{IPVForensicResult}(x, ...)
}
\description{
Print Method for IPV Forensic Result
}
