% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{print_validation_report}
\alias{print_validation_report}
\title{Print Validation Report}
\usage{
print_validation_report(metrics)
}
\arguments{
\item{metrics}{Metrics list from calculate_metrics}
}
\description{
Print Validation Report
}
