% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_nvdrs_data}
\alias{read_nvdrs_data}
\title{Read NVDRS Data}
\usage{
read_nvdrs_data(file_path)
}
\arguments{
\item{file_path}{Path to CSV file}
}
\value{
Data frame with narratives
}
\description{
Read NVDRS Data
}
