% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{split_into_batches}
\alias{split_into_batches}
\title{Split Data into Batches}
\usage{
split_into_batches(data, batch_size = 50)
}
\arguments{
\item{data}{Data frame}

\item{batch_size}{Size of each batch}
}
\value{
List of data frames
}
\description{
Split Data into Batches
}
