% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv_detection.R
\name{validate_results}
\alias{validate_results}
\title{Validate Results}
\usage{
validate_results(results)
}
\arguments{
\item{results}{Data frame with predictions and ManualIPVFlag}
}
\value{
Validation metrics
}
\description{
Validate Results
}
