% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{calculate_agreement}
\alias{calculate_agreement}
\title{Calculate Agreement}
\usage{
calculate_agreement(results)
}
\arguments{
\item{results}{Data frame with le_ipv and cme_ipv columns}
}
\value{
Agreement statistics
}
\description{
Calculate Agreement
}
