% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate Metrics}
\usage{
calculate_metrics(predictions, actual = NULL)
}
\arguments{
\item{predictions}{Data frame with ipv_detected column}

\item{actual}{Data frame with ManualIPVFlag column}
}
\value{
List with metrics
}
\description{
Calculate Metrics
}
