% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Generate Confusion Matrix (Modernized)}
\usage{
confusion_matrix(predictions, actual)
}
\arguments{
\item{predictions}{Predictions vector or data frame}

\item{actual}{Actual values vector or data frame}
}
\value{
Confusion matrix as tibble
}
\description{
Generate Confusion Matrix (Modernized)
}
