% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv_detection.R
\name{detect_ipv}
\alias{detect_ipv}
\title{Detect IPV in Narrative}
\usage{
detect_ipv(narrative, type = "LE", config, conn = NULL)
}
\arguments{
\item{narrative}{Narrative text}

\item{type}{"LE" or "CME"}

\item{config}{Configuration}

\item{conn}{Database connection}
}
\value{
Detection result
}
\description{
Detect IPV in Narrative
}
