% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv_detection.R
\name{load_config}
\alias{load_config}
\title{Load Configuration}
\usage{
load_config(config_path = NULL)
}
\arguments{
\item{config_path}{Path to settings.yml}
}
\value{
Configuration list
}
\description{
Load Configuration
}
