% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{validate_input_data}
\alias{validate_input_data}
\title{Validate Input Data}
\usage{
validate_input_data(data)
}
\arguments{
\item{data}{Data frame to validate}
}
\value{
Validated data frame
}
\description{
Validate Input Data
}
